Add server side validation to ClientController

+++ TP1/src/main/java/com/bcmworld/tp1/controller/ClientController.java
@@ -3,12 +3,18 @@ package com.bcmworld.tp1.controller;
 import com.bcmworld.tp1.model.daos.GenericDAO;
 import com.bcmworld.tp1.model.dtos.ClientDTO;
 import com.google.gson.Gson;
+import com.google.gson.reflect.TypeToken;
+import org.apache.commons.validator.routines.DoubleValidator;
+import org.apache.commons.validator.routines.EmailValidator;
+import org.apache.commons.validator.routines.RegexValidator;
 import spark.ModelAndView;
 import spark.Route;
 import spark.TemplateViewRoute;
 import spark.template.velocity.VelocityTemplateEngine;
 
-import java.util.*;
+import java.lang.reflect.Type;
+import java.util.HashMap;
+import java.util.Map;
 
 import static spark.Spark.*;
 
@@ -31,8 +37,7 @@ public class ClientController {
         return (request, response) -> {
             Map<String, Object> model = new HashMap<>();
             int index = Integer.valueOf(request.params(":index"));
-//            clientDao.findAll(index * CLIENTS_PER_PAGE, CLIENTS_PER_PAGE).forEach(clientDTO -> System.out.println(clientDTO.getName()));
-            model.put("clients", clientDao.findAll((index-1) * CLIENTS_PER_PAGE, CLIENTS_PER_PAGE));
+            model.put("clients", clientDao.findAll((index - 1) * CLIENTS_PER_PAGE, CLIENTS_PER_PAGE));
             model.put("index", index);
             model.put("pages", Math.floor(clientDao.countAll() / CLIENTS_PER_PAGE) + 1);
             return new ModelAndView(model, "public/velocity/clients.vm");
@@ -64,16 +69,18 @@ public class ClientController {
 
         return (request, response) -> {
 
-            // TODO: validation and 400 status code in case of error
-            //System.out.println(client.getName() == null);
-
             String body = request.body();
-            ClientDTO client = new Gson().fromJson(body, ClientDTO.class);
 
+            String invalid = validateClientJson(body);
+            if (invalid != null) {
+                response.status(400);
+                return invalid;
+            }
+
+            ClientDTO client = new Gson().fromJson(body, ClientDTO.class);
             clientDao.save(client);
 
             response.status(201);
-
             return "{ \"result\" : \"OK\"}";
         };
     }
@@ -82,16 +89,18 @@ public class ClientController {
 
         return (request, response) -> {
 
-            // TODO: validation and 400 status code in case of error
-            //System.out.println(client.getName() == null);
-
             String body = request.body();
-            ClientDTO client = new Gson().fromJson(body, ClientDTO.class);
 
+            String invalid = validateClientJson(body);
+            if (invalid != null) {
+                response.status(400);
+                return invalid;
+            }
+
+            ClientDTO client = new Gson().fromJson(body, ClientDTO.class);
             clientDao.update(client);
 
             response.status(201);
-
             return "{ \"result\" : \"OK\"}";
         };
     }
@@ -100,14 +109,78 @@ public class ClientController {
 
         return (request, response) -> {
 
-            // TODO: validation and 400
-
             String body = request.body();
-            ClientDTO client = new Gson().fromJson(body, ClientDTO.class);
+            Type type = new TypeToken<Map<String, String>>() {
+            }.getType();
+            Map<String, String> myMap = new Gson().fromJson(body, type);
+            if (!new RegexValidator("^[0-9]{1,}$").isValid(myMap.get("cuitDNI"))) {
+                return "{ 'result':'error', 'invalid':'cuitDNI'";
+            }
 
+            ClientDTO client = new Gson().fromJson(body, ClientDTO.class);
             if (client.getCuitDNI() != null) clientDao.delete(client.getCuitDNI());
 
+            response.status(201);
             return "{ \"result\" : \"OK\"}";
         };
+    }
+
+    private String validateClientJson(String json) {
+
+        Type type = new TypeToken<Map<String, String>>() {
+        }.getType();
+        Map<String, String> myMap = new Gson().fromJson(json, type);
+
+        boolean error = false;
+        String invalid = "{ 'result':'error', 'invalid':[";
+
+        if (!new RegexValidator("^[0-9]{1,}$", false).isValid(myMap.get("cuitDNI"))) {
+            error = true;
+            invalid += "'cuitDNI', ";
+        }
+        if (myMap.get("name") == null || myMap.get("name").isEmpty()) {
+            error = true;
+            invalid += "'name', ";
+        }
+        if (myMap.get("surname") == null || myMap.get("surname").isEmpty()) {
+            error = true;
+            invalid += "'surname', ";
+        }
+        if (myMap.get("legalName") == null || myMap.get("legalName").isEmpty()) {
+            error = true;
+            invalid += "'legalName', ";
+        }
+        if (myMap.get("phone") == null || myMap.get("phone").isEmpty()) {
+            error = true;
+            invalid += "'phone', ";
+        }
+        if (!EmailValidator.getInstance().isValid(myMap.get("mail"))) {
+            error = true;
+            invalid += "'mail', ";
+        }
+        if (!DoubleValidator.getInstance().isValid(myMap.get("longitude"))) {
+            error = true;
+            invalid += "'longitude', ";
+        }
+        if (!DoubleValidator.getInstance().isValid(myMap.get("latitude"))) {
+            error = true;
+            invalid += "'latitude', ";
+        }
+        if (myMap.get("address") == null || myMap.get("address").isEmpty()) {
+            error = true;
+            invalid += "'address', ";
+        }
+        if (myMap.get("type") == null) {
+            error = true;
+            invalid += "'type', ";
+        } else if (!(myMap.get("type").equalsIgnoreCase("Client")
+                || myMap.get("type").equalsIgnoreCase("Contact"))) {
+            error = true;
+            invalid += "'type', ";
+        }
+        invalid += "]}";
+
+        if (error) return invalid;
+        else return null;
     }
 }
